test: remove duplication
